Move out DoKeysOverlap and splitwstring
