+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/KeyboardManagerEditorLibrary.vcxproj
@@ -36,6 +36,7 @@
     <ClInclude Include="BufferValidationHelpers.h" />
     <ClInclude Include="Dialog.h" />
     <ClInclude Include="EditKeyboardWindow.h" />
+    <ClInclude Include="EditorHelpers.h" />
     <ClInclude Include="EditShortcutsWindow.h" />
     <ClInclude Include="KeyboardManagerEditorStrings.h" />
     <ClInclude Include="KeyboardManagerState.h" />
@@ -54,6 +55,7 @@
     <ClCompile Include="BufferValidationHelpers.cpp" />
     <ClCompile Include="Dialog.cpp" />
     <ClCompile Include="EditKeyboardWindow.cpp" />
+    <ClCompile Include="EditorHelpers.cpp" />
     <ClCompile Include="EditShortcutsWindow.cpp" />
     <ClCompile Include="KeyboardManagerEditorStrings.cpp" />
     <ClCompile Include="KeyboardManagerState.cpp" />
