+++ src/modules/keyboardmanager/KeyboardManagerEditorTest/KeyboardManagerEditorTest.vcxproj
@@ -44,6 +44,7 @@
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(CIBuild)'!='true'">Create</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="EditorHelpersTests.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="MockedInput.h" />
