+++ src/modules/keyboardmanager/common/Shortcut.h
@@ -21,6 +21,9 @@ private:
     ModifierKey shiftKey;
     DWORD actionKey;
 
+    // Function to split a wstring based on a delimiter and return a vector of split strings
+    std::vector<std::wstring> splitwstring(const std::wstring& input, wchar_t delimiter);
+
 public:
     // By default create an empty shortcut
     Shortcut() :
